+++ pkgs/shelf_router/test/router_test.dart
@@ -40,8 +40,8 @@ void main() {
 
   Future<String> get(String path) =>
       http.read(Uri.parse(server.url.toString() + path));
-  Future<int> head(String path) async =>
-      (await http.head(Uri.parse(server.url.toString() + path))).statusCode;
+  Future<http.Response> head(String path) =>
+      http.head(Uri.parse(server.url.toString() + path));
 
   test('get sync/async handler', () async {
     var app = Router();
@@ -67,9 +67,20 @@ void main() {
     expect(await get('/async-hello'), 'hello-world');
     expect(await get('/wrong-path'), 'not-found');
 
-    expect(await head('/sync-hello'), 200);
-    expect(await head('/async-hello'), 200);
-    expect(await head('/wrong-path'), 200);
+    final syncHead = await head('/sync-hello');
+    expect(syncHead.statusCode, 200);
+    expect(syncHead.body, isEmpty);
+    expect(syncHead.headers['content-length'], '11');
+
+    final asyncHead = await head('/async-hello');
+    expect(asyncHead.statusCode, 200);
+    expect(asyncHead.body, isEmpty);
+    expect(asyncHead.headers['content-length'], '11');
+
+    final wrongPathHead = await head('/wrong-path');
+    expect(wrongPathHead.statusCode, 200);
+    expect(wrongPathHead.body, isEmpty);
+    expect(wrongPathHead.headers['content-length'], '9');
   });
 
   test('params', () async {